Remove Python 2 shims

+++ openpyxl/worksheet/_reader.py
@@ -4,7 +4,6 @@
 from warnings import warn
 
 # compatibility imports
-from openpyxl.compat import long
 from openpyxl.xml.functions import iterparse
 
 # package imports
@@ -77,7 +76,7 @@ def _cast_number(value):
     "Convert numbers as string to an int or float"
     if "." in value or "E" in value or "e" in value:
         return float(value)
-    return long(value)
+    return int(value)
 
 
 class WorkSheetParser(object):
